chore(build): standardize package builds on one shipped-source rule
Shared collectBuildEntries() + BUILD_EXCLUDE_GLOBS in @mionjs/devtools
replace ten copy-pasted collectors: everything under src/ ships unless
suffixed .spec/.test/.bench/.stub. Every built package now has a uniform
tsconfig.build.json read by BOTH the dts plugin and the runtypes plugin,
so the build-time type scan no longer includes test files — core's build
genDir drops to its 2 error-class entries, router's from 1386 modules to
143, and router stops publishing its .bench files (which dragged ~120
generated cache modules into the dist).
