+++ packages/core/tsconfig.build.json
@@ -1,5 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "references": [{"path": "../core"}],
   "exclude": ["node_modules", ".dist", "build", "*.config.ts", "globalSetup.ts", "buildTestBundle.ts", "**/*.spec.ts", "**/*.test.ts", "**/*.bench.ts", "**/*.stub.ts"]
 }
