+++ packages/core/vite.config.ts
@@ -1,35 +1,10 @@
 import {defineConfig} from 'vite';
 import {resolve} from 'path';
-import {readdirSync, statSync} from 'fs';
 import dts from 'vite-plugin-dts';
-import {mionVitePlugin, cjsPackageJsonPlugin} from '@mionjs/devtools/vite-plugin';
+import {mionVitePlugin, cjsPackageJsonPlugin, collectBuildEntries, BUILD_EXCLUDE_GLOBS} from '@mionjs/devtools/vite-plugin';
 
-// Get all TypeScript files from src directory (excluding spec/test files)
-function getSourceFiles(dir: string, base = ''): Record<string, string> {
-  const entries: Record<string, string> = {};
-  const files = readdirSync(dir);
-
-  for (const file of files) {
-    const fullPath = resolve(dir, file);
-    const relativePath = base ? `${base}/${file}` : file;
-
-    if (statSync(fullPath).isDirectory()) {
-      Object.assign(entries, getSourceFiles(fullPath, relativePath));
-    } else if (file.endsWith('.ts') && !file.endsWith('.spec.ts') && !file.endsWith('.test.ts')) {
-      const name = relativePath.replace(/\.ts$/, '');
-      entries[name] = fullPath;
-    }
-  }
-
-  return entries;
-}
-
-// Build entry points: index.ts + all src files
-const srcEntries = getSourceFiles(resolve(__dirname, 'src'));
-const entry: Record<string, string> = {
-  index: resolve(__dirname, 'index.ts'),
-  ...Object.fromEntries(Object.entries(srcEntries).map(([name, path]) => [`src/${name}`, path])),
-};
+// Build entry points: index.ts + all shippable src files (shared rule in @mionjs/devtools)
+const entry = collectBuildEntries(__dirname);
 
 export default defineConfig({
   esbuild: {
@@ -39,14 +14,15 @@ export default defineConfig({
     cjsPackageJsonPlugin('.dist/cjs'),
     mionVitePlugin({
       runTypes: {
-        tsConfig: resolve(__dirname, 'tsconfig.json'),
+        tsConfig: resolve(__dirname, 'tsconfig.build.json'),
       },
     }),
     dts({
       outDir: ['.dist/cjs', '.dist/esm'],
       include: ['index.ts', 'src/**/*.ts'],
-      exclude: ['**/*.spec.ts', '**/*.test.ts'],
+      exclude: [...BUILD_EXCLUDE_GLOBS],
       pathsToAliases: false,
+      tsconfigPath: resolve(__dirname, 'tsconfig.build.json'),
     }),
   ],
   build: {
