+++ packages/drizzle/src/stubs-formats-mappings/common.stub.ts
